All pages updated except live and payment history

+++ src/main/java/Vehicle/example/Management/Repository/UserRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepo extends JpaRepository<UserList,Integer> {
+public interface UserRepo extends JpaRepository<UserList,Long> {
     UserList findByusernameAndPassword(String username, String password);
     Optional<UserList> findByUsername(String username);
 }
